+++ WizardPanels/tests/WizardPanels-specs.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {mount, shallow} from 'enzyme';
 
-import {View, WizardPanel, WizardPanelBase} from '../WizardPanel';
+import {Panel, WizardPanels, WizardPanelsBase} from '../';
 
 describe('WizardPanel Specs', () => {
 	test(
@@ -10,7 +10,7 @@ describe('WizardPanel Specs', () => {
 			const title = 'WizardPanel title';
 
 			const wizardPanel = shallow(
-				<WizardPanelBase title={title} />
+				<WizardPanelsBase title={title} />
 			);
 
 			const headerTitle = wizardPanel.find({type: 'wizard'}).prop('title');
@@ -29,9 +29,9 @@ describe('WizardPanel Specs', () => {
 			const viewSubtitle = 'View subtitle';
 
 			const wizardPanel = mount(
-				<WizardPanel subtitle={wizardSubtitle}>
-					<View subtitle={viewSubtitle} />
-				</WizardPanel>
+				<WizardPanels subtitle={wizardSubtitle}>
+					<Panel subtitle={viewSubtitle} />
+				</WizardPanels>
 			);
 
 			const headerSubtitle = wizardPanel.find('Header').prop('subtitle');
@@ -50,9 +50,9 @@ describe('WizardPanel Specs', () => {
 			const viewTitle = 'View title';
 
 			const wizardPanel = mount(
-				<WizardPanel title={wizardTitle}>
-					<View title={viewTitle} />
-				</WizardPanel>
+				<WizardPanels title={wizardTitle}>
+					<Panel title={viewTitle} />
+				</WizardPanels>
 			);
 
 			const headerTitle = wizardPanel.find('Header').prop('title');
@@ -70,9 +70,9 @@ describe('WizardPanel Specs', () => {
 			const viewSubtitle = 'View subtitle';
 
 			const wizardPanel = mount(
-				<WizardPanel>
-					<View subtitle={viewSubtitle} />
-				</WizardPanel>
+				<WizardPanels>
+					<Panel subtitle={viewSubtitle} />
+				</WizardPanels>
 			);
 
 			const headerSubtitle = wizardPanel.find('Cell.titleCell .text').last().text();
@@ -90,9 +90,9 @@ describe('WizardPanel Specs', () => {
 			const viewFooter = 'View footer';
 
 			const wizardPanel = mount(
-				<WizardPanel>
-					<View footer={viewFooter} />
-				</WizardPanel>
+				<WizardPanels>
+					<Panel footer={viewFooter} />
+				</WizardPanels>
 			);
 
 			const footerText = wizardPanel.find('.footer').text();
@@ -108,14 +108,14 @@ describe('WizardPanel Specs', () => {
 		'should have View buttons rendered in `.buttonContainer`',
 		() => {
 			const wizardPanel = mount(
-				<WizardPanel>
-					<View>
+				<WizardPanels>
+					<Panel>
 						<buttons>
 							<button>Button 1</button>
 							<button>Button 2</button>
 						</buttons>
-					</View>
-				</WizardPanel>
+					</Panel>
+				</WizardPanels>
 			);
 
 			const buttons = wizardPanel.find('.buttonContainer').find('button').length;
@@ -133,11 +133,11 @@ describe('WizardPanel Specs', () => {
 			const contentText = 'content';
 
 			const wizardPanel = mount(
-				<WizardPanel>
-					<View>
+				<WizardPanels>
+					<Panel>
 						{contentText}
-					</View>
-				</WizardPanel>
+					</Panel>
+				</WizardPanels>
 			);
 
 			const content = wizardPanel.find('.content').find('.enact-fit').text();
@@ -155,7 +155,7 @@ describe('WizardPanel Specs', () => {
 			const nextButtonText = 'next';
 
 			const wizardPanel = shallow(
-				<WizardPanelBase total={2} nextButtonText={nextButtonText} />
+				<WizardPanelsBase total={2} nextButtonText={nextButtonText} />
 			);
 
 			// Using slot as a proxy to find Button since it's name isn't set
@@ -174,7 +174,7 @@ describe('WizardPanel Specs', () => {
 			const prevButtonText = 'previous';
 
 			const wizardPanel = shallow(
-				<WizardPanelBase index={1} prevButtonText={prevButtonText} />
+				<WizardPanelsBase index={1} prevButtonText={prevButtonText} />
 			);
 
 			const prevButton = wizardPanel.find({slot: 'slotBefore'});
@@ -190,7 +190,7 @@ describe('WizardPanel Specs', () => {
 		'should hide next button on the last view',
 		() => {
 			const wizardPanel = shallow(
-				<WizardPanelBase index={2} total={3} />
+				<WizardPanelsBase index={2} total={3} />
 			);
 
 			const nextButton = wizardPanel.find({slot: 'slotAfter'});
@@ -206,7 +206,7 @@ describe('WizardPanel Specs', () => {
 		'should hide previous button on the first view',
 		() => {
 			const wizardPanel = shallow(
-				<WizardPanelBase index={0} total={3} />
+				<WizardPanelsBase index={0} total={3} />
 			);
 
 			const prevButton = wizardPanel.find({slot: 'slotBefore'});
@@ -222,7 +222,7 @@ describe('WizardPanel Specs', () => {
 		'should hide next button with `noNextButton`',
 		() => {
 			const wizardPanel = shallow(
-				<WizardPanelBase index={2} noNextButton total={4} />
+				<WizardPanelsBase index={2} noNextButton total={4} />
 			);
 
 			const nextButton = wizardPanel.find({slot: 'slotAfter'});
@@ -238,7 +238,7 @@ describe('WizardPanel Specs', () => {
 		'should hide previous button with `noPrevButton`',
 		() => {
 			const wizardPanel = shallow(
-				<WizardPanelBase index={2} noPrevButton total={4} />
+				<WizardPanelsBase index={2} noPrevButton total={4} />
 			);
 
 			const prevButton = wizardPanel.find({slot: 'slotBefore'});
@@ -256,7 +256,7 @@ describe('WizardPanel Specs', () => {
 		() => {
 			const index = 1;
 			const wizardPanel = shallow(
-				<WizardPanelBase index={index} total={5} />
+				<WizardPanelsBase index={index} total={5} />
 			);
 
 			const expected = {current: index + 1};
@@ -272,9 +272,9 @@ describe('WizardPanel Specs', () => {
 			// FIXME: Temporary selector until our components have corrected display names
 			const viewManager = '.content > *';
 			const wizardPanel = shallow(
-				<WizardPanelBase>
-					<View />
-				</WizardPanelBase>
+				<WizardPanelsBase>
+					<Panel />
+				</WizardPanelsBase>
 			);
 
 			let actual = wizardPanel.find(viewManager).prop('noAnimation');